Add some validation so that the redis provider can also remove plain openshift deployments.

+++ web/modules/custom/shp_redis_support/src/Plugin/OrchestrationProvider/OpenShiftWithRedis.php
@@ -154,8 +154,12 @@ class OpenShiftWithRedis extends OpenShiftOrchestrationProvider {
    */
   public function deleteRedisDeployment(string $deployment_name) {
     $redis_name = $deployment_name . '-redis';
-    $this->client->deleteService($redis_name);
-    $this->client->deleteDeploymentConfig($redis_name);
+    if ($this->client->getService($redis_name)) {
+      $this->client->deleteService($redis_name);
+    }
+    if ($this->client->getDeploymentConfig($redis_name)) {
+      $this->client->deleteDeploymentConfig($redis_name);
+    }
   }
 
 }
